requirements.txt
aiohttp==3.7.4.post0
ansible==3.3.0
ansible-core==2.11.0
apparmor==3.0.1
appdirs==1.4.4
application-utility==1.3.2
arandr==0.1.10
asn1crypto==1.4.0
astroid==2.6.6
async-timeout==3.0.1
attrs==20.3.0
Automat==20.2.0
bauh==0.9.16
bcrypt==3.2.0
Beaker==1.11.0
btrfsutil==5.11.1
CacheControl==0.12.6
cached-property==1.5.2
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
constantly==15.1.0
contextlib2==0.6.0.post1
cryptography==3.4.7
cupshelpers==1.0
deluge==2.0.4.dev56
discord.py==1.7.3
distlib==0.3.1
distro==1.5.0
dnspython==1.16.0
docker==5.0.0
docker-compose==1.29.1
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
Flask==1.1.2
GeoIP==1.3.2
gufw==21.4.0
html5lib==1.1
hyperlink==21.0.0
idna==3.1
importlib-metadata==4.0.1
incremental==21.3.0
isc==2.0
isort==5.9.3
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
lazy-object-proxy==1.6.0
lensfun==0.3.95
LibAppArmor==3.0.1
louis==3.17.0
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
meson==0.58.0
more-itertools==8.7.0
msgpack==1.0.2
multidict==5.1.0
mysql-connector-python==8.0.26
netsnmp-python==1.0a1
npyscreen==4.10.5
numpy==1.20.2
ordered-set==4.0.2
packaging==20.9
pacman-mirrors==4.21.2
paramiko==2.7.2
pep517==0.10.0
Pillow==8.2.0
ply==3.11
progress==1.5
protobuf==3.17.3
psutil==5.8.0
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.0
pycparser==2.20
pycups==2.0.1
pycurl==7.43.0.6
PyGObject==3.40.1
PyHamcrest==1.10.0
pylint==2.9.6
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-sip==12.8.1
pyrsistent==0.17.3
pysmbc==1.0.23
python-dateutil==2.8.1
python-dotenv==0.17.1
python-libtorrent==1.2.11
python-xapp==2.0.2
python-xlib==0.29
pytoml==0.1.21
pyxdg==0.27
PyYAML==5.4.1
ranger-fm==1.9.3
rencode==1.0.6
reportlab==3.5.66
requests==2.25.1
resolvelib==0.5.4
retrying==1.3.3
scipy==1.6.3
service-identity==18.1.0
setproctitle==1.2.2
sip==4.19.25
six==1.15.0
speedtest-cli==2.1.3
team==1.0
texttable==1.6.3
TLPUI==1.3.1.post8
toml==0.10.2
Twisted==21.2.0
typing-extensions==3.10.0.0
ueberzug==18.1.9
ufw==0.36
urllib3==1.26.4
urwid==2.1.2
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
zope.interface==5.4.0
